+++ libEnigma/src/x86/scoreAvx.c
@@ -15,10 +15,10 @@
 #include "score_inlines.h"
 
 // SSSE3 scores
-static double icscoreAvx( const Key* const restrict key, scoreLength_t len );
-static int   uniscoreAvx( const Key* const restrict key, scoreLength_t len );
-static int    biscoreAvx( const Key* const restrict key, scoreLength_t len );
-static int   triscoreAvx( const Key* const restrict key, scoreLength_t len );
+static uint16_t icscoreAvx( const Key* const restrict key, scoreLength_t len );
+static int     uniscoreAvx( const Key* const restrict key, scoreLength_t len );
+static int      biscoreAvx( const Key* const restrict key, scoreLength_t len );
+static int     triscoreAvx( const Key* const restrict key, scoreLength_t len );
 
 enigma_score_function_t enigmaScoreAvx = { triscoreAvx,  biscoreAvx , icscoreAvx,  uniscoreAvx } ;
 
@@ -26,7 +26,7 @@ union ScoringDecodedMessage decodedMsgPartAvx;
 
 __attribute__ ((flatten))
 __attribute__ ((optimize("unroll-loops")))
-static double icscoreAvx( const Key* const restrict key, scoreLength_t len ) {
+static uint16_t icscoreAvx( const Key* const restrict key, scoreLength_t len ) {
     DecodeScoredMessagePartSsse3( key, len, &decodedMsgPartAvx );
     return ComputeIcscoreFromDecodedMsgSsse3( &decodedMsgPartAvx, len );
 }
